changed challangeRating type

+++ src/main/java/org/example/Monster.java
@@ -3,7 +3,7 @@ package org.example;
 public class Monster {
     private String type;
     private String name;
-    private Integer challangeRating;
+    private double challangeRating;
     private Integer health;
     private Integer maxHealth;
 }
